amended based on comments
