+++ backend/tests/rep_sfx_test.py
@@ -15,7 +15,7 @@ from backend.object import Object
 from backend.domain import Domain
 from backend.state import State
 from backend.action import Action
-from backend.special_effect import RepetitiveEffect
+from backend.special_effects.repetitive_effect import RepetitiveEffect
 
 is_lettuce_def = Predicate().initialize("is_lettuce", ["item"])
 is_table_def = Predicate().initialize("is_table", ["station"])
